+++ examples/use_api/mlw_tree.ml
@@ -20,8 +20,6 @@ API calls
 open Why3
 let config : Whyconf.config = Whyconf.init_config None
 let main : Whyconf.main = Whyconf.get_main config
-let libdir = Whyconf.libdir main
-let datadir = Whyconf.datadir main
 let env : Env.env = Env.create_env (Whyconf.loadpath main)
 open Ptree
 open Ptree_helpers
@@ -358,7 +356,7 @@ let alt_ergo : Whyconf.config_prover =
 
 let alt_ergo_driver : Driver.driver =
   try
-    Whyconf.load_driver main env alt_ergo
+    Driver.load_driver main env alt_ergo
   with e ->
     eprintf "Failed to load driver for alt-ergo: %a@."
       Exn_printer.exn_printer e;
@@ -368,7 +366,7 @@ let () =
   List.iteri (fun i t ->
       let call =
         Driver.prove_task ~limit:Call_provers.empty_limit
-          ~libdir ~datadir
+          ~config:main
           ~command:alt_ergo.Whyconf.command alt_ergo_driver t in
       let r = Call_provers.wait_on_call call in
       printf "@[On task %d, alt-ergo answers %a@." (succ i)