+++ src/mlw/vc.ml
@@ -570,9 +570,9 @@ let add_loc_attr label loc attrs =
   match loc with
   | None -> attrs
   | Some loc ->
-      let filename, line, bchar, echar = Loc.get loc in
+      let filename, bline, bchar, eline, echar = Loc.get loc in
       let attr = Format.kasprintf Ident.create_attribute
-          "%s:%s:%d:%d:%d" label filename line bchar echar in
+          "%s:%s:%d:%d:%d:%d" label filename bline bchar eline echar in
       Sattr.add attr attrs
 
 (* translate the expression [e] into a k-expression:
@@ -653,7 +653,7 @@ let rec k_expr env lps e res xmap =
         (* [ VC(ce) (if ce is a lambda executed in-place)
            | STOP pre
            | HAVOC ; [ ASSUME post | ASSUME xpost ; RAISE ] ] *)
-       (* Format.eprintf "[Vc.term_of_post/Eexec] res = %a@." print_pv_attr res; *)
+       (* Format.eprintf "[Vc.term_of_post/Eexec] e = %a@." print_expr e; *)
         let p, (oldies, sbs) = match pre with
           (* for recursive calls, compute the 'variant decrease'
              precondition and rename the oldies to avoid clash *)
@@ -663,16 +663,32 @@ let rec k_expr env lps e res xmap =
               let d = decrease env loc attrs expl_variant ovl nvl in
               wp_and d (wp_of_pre loc attrs pl), renew_oldies oldies
           | pl -> wp_of_pre loc attrs pl, (oldies, Mvs.empty) in
+        let is_fully_applied = ce.c_cty.cty_args = [] in
+        let is_constructor_or_projection ls =
+          Typeinv.is_trusted_constructor env.known_map ls ||
+            Typeinv.is_trusted_projection env.known_map ls e.e_ity
+        in
         let trusted = match ce.c_node with
           | (Capp ({rs_logic = RLls ls}, _) | Cpur (ls, _))
-               when ce.c_cty.cty_args = [] (* fully applied *) ->
-              Typeinv.is_trusted_constructor env.known_map ls ||
-              Typeinv.is_trusted_projection env.known_map ls e.e_ity
+               when is_fully_applied -> is_constructor_or_projection ls
           | _ -> false in
         let rds = cty.cty_effect.eff_reads in
         let aff = pvs_affected cty.cty_effect.eff_covers rds in
         let pinv = if trusted then [] else inv_of_pvs env e.e_loc rds in
         let qinv = if trusted then [] else inv_of_pvs env e.e_loc aff in
+        let need_trace = match ce.c_node with
+          | Capp ({rs_logic = RLls ls}, args) (* `let function` or `val function` *)
+            | Cpur (ls, args) (* direct application of a logic symbol *)
+            ->
+             let is_a_constant = args = [] in
+             not is_fully_applied ||
+               not (is_a_constant || is_constructor_or_projection ls)
+          | Capp ({rs_logic = (RLnone|RLlemma|RLpv _)}, _)
+            -> true
+          | Cfun _ -> true
+          | Cany -> true
+        in
+        (* Format.eprintf "[Vc.term_of_post/Eexec] need_trace = %b@." need_trace; *)
         let k_of_post expl v ql =
           let k v =
             let sp = sp_of_post loc attrs expl v ql in
@@ -683,13 +699,6 @@ let rec k_expr env lps e res xmap =
             | Some (t, sp) ->
                Klet (v, t_tag t, List.fold_right sp_and rinv sp)
             | None ->  Kval ([v], List.fold_right sp_and rinv sp) in
-          let need_trace = match ce.c_node with
-            | (Capp ({rs_logic = RLls _ls}, _) (* `let function` or `val function` *)
-               | Cpur (_ls, _)) (* direct application of a logic symbol *)
-              ->
-               ce.c_cty.cty_args <> [] (* unless not fully applied *)
-          | _ -> true
-            in
           if env.keep_trace && need_trace then
             let vv = explicit_result loc e.e_attrs ce v.pv_ity in
             Kseq(k v,0,Klet(vv, t_var v.pv_vs, t_true))